refactor: split DataCards into separate card components
