+++ Frontend/mobile_expo/app/components/chatbot/cards/index.ts
@@ -1,0 +1,5 @@
+export { default as ClientCard } from './ClientCard';
+export { default as ProjectCard } from './ProjectCard';
+export { default as DocumentCard } from './DocumentCard';
+export { default as ScheduleCard } from './ScheduleCard';
+export { default as StaffCard } from './StaffCard';
